feat: use mapParallel2 for interaction partners to speed up response time
